requirements.txt
setuptools>=65.0.0
wheel>=0.40.0
python-telegram-bot[job-queue]==22.4
apscheduler==3.10.4
gspread==6.1.2
google-auth==2.34.0
google-auth-oauthlib==1.2.2
requests==2.32.5
httpx==0.27.2
python-dotenv==1.0.1
deep-translator==1.10.1
openai==1.0.0
pydantic==2.11.9
uvicorn==0.30.0
fastapi==0.111.0
sqlalchemy==2.1.0
aiohttp==3.9.5
